% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_map_state}
\alias{fars_map_state}
\title{\code{fars_map_state} : This function takes the state number and the vector of years as input.
It will then find the relevant csv files using the \code{make_filename} function and will
read those csv files using the  \code{fars_read} function. If the state number is not found
it will display an error. Otherwise, it will filter the dataset, filtering the data with the
given state number. It then plots the points where the accidents have occured in that state.
If no accidents occured, a message is displayed.}
\usage{
fars_map_state(state.num, year)
}
\value{
a data.frame of filtered data, if there are no rows then returns invisible(NULL)
}
\description{
\code{fars_map_state} : This function takes the state number and the vector of years as input.
It will then find the relevant csv files using the \code{make_filename} function and will
read those csv files using the  \code{fars_read} function. If the state number is not found
it will display an error. Otherwise, it will filter the dataset, filtering the data with the
given state number. It then plots the points where the accidents have occured in that state.
If no accidents occured, a message is displayed.
}
