% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_read}
\alias{fars_read}
\title{-------------------------------------------------------------
COURSERA : BUILDING R PACKAGES
PEER REVIEW ASSIGNMENT
FILE : fars_function.R
SUBMITTED BY : JUNAID KHAN}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that takes the name of the
file to read}
}
\value{
A file that has read the csv data or an error stating
 that the data file does not exist
}
\description{
-------------------------------------------------------------
}
\details{
\code{fars_read} : Read fars data
This function reads data from the US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System.
The data provides us census regarding the American public yearly data
regarding fatal injuries sufferred in motor vehicle traffic crashes.

For more information, see:
\itemize{
  \item{\url{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}}
  \item{\url{https://en.wikipedia.org/wiki/Fatality_Analysis_Reporting_System}}
}
}
