% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{\code{fars_summarize_years} : The function reads the years as a numeric input.
The function will read all the csv files found in the directory, using the \code{fars_read_years}
function and bind them row-wise. It will then group the data based on year and
MONTH and summarize the data.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of year to be searched}
}
\value{
A dataset which is a Summary of all the relevant datasets in the directory,
binded row-wise, grouped-by the year and MONTH
}
\description{
\code{fars_summarize_years} : The function reads the years as a numeric input.
The function will read all the csv files found in the directory, using the \code{fars_read_years}
function and bind them row-wise. It will then group the data based on year and
MONTH and summarize the data.
}
