% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_function.R
\name{make_filename}
\alias{make_filename}
\title{\code{make_filename} : Read the input year and find the relevant
dataset for that year. If the approriate year is not written, the
sprintf function will return an error.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numeric value of the year, used to find the relevant dataset}
}
\value{
A character vector filename
}
\description{
\code{make_filename} : Read the input year and find the relevant
dataset for that year. If the approriate year is not written, the
sprintf function will return an error.
}
